Unificater-Backend/requirements.txt
aniso8601==8.0.0
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.4.2
asttokens==2.0.4
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.2
django-cors-headers==3.5.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.12.2
docutils==0.16
django-extensions==3.1.0
entrypoints==0.3
idna==2.6
isort==5.6.4
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mongoengine==0.20.0
mypy==0.790
mypy-extensions==0.4.3
mysql-connector==2.2.9
mysqlclient==2.0.2
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
numpy==1.19.2
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.3
parso==0.7.1
pickleshare==0.7.5
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.14.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
Pygments==2.7.2
pylint==2.6.0
pymongo==3.11.0
pyodbc==4.0.30
pyparsing==2.4.7
pyrsistent==0.17.3
pyserial==3.5
python-dateutil==2.8.1
pytz==2020.4
pywin32==300
pywinpty==0.5.7
requests==2.18.4
Send2Trash==1.5.0
six==1.15.0
SQLAlchemy==1.3.19
sqlparse==0.4.1
terminado==0.9.1
testpath==0.4.4
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typed-ast==1.4.1
typing-extensions==3.7.4.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.16.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==1.2.0